Assignment 3

+++ routes/dishRouter.js
@@ -22,7 +22,7 @@ dishRouter
       )
       .catch((err) => next(err));
   })
-  .post(authenticate.verifyUser ,(req, res, next) => {
+  .post(authenticate.verifyUser, authenticate.verifyAdmin,(req, res, next) => {
     Dishes.create(req.body)
       .then(
         (dish) => {
@@ -35,11 +35,11 @@ dishRouter
       )
       .catch((err) => next(err));
   })
-  .put(authenticate.verifyUser,(req, res, next) => {
+  .put(authenticate.verifyUser, authenticate.verifyAdmin,(req, res, next) => {
     res.statusCode = 403;
     res.end("PUT Operation not supported on /dishes ");
   })
-  .delete(authenticate.verifyUser,(req, res, next) => {
+  .delete(authenticate.verifyUser, authenticate.verifyAdmin,(req, res, next) => {
     Dishes.deleteMany({})
       .then(
         (resp) => {
@@ -67,11 +67,11 @@ dishRouter
       )
       .catch((err) => next(err));
   })
-  .post(authenticate.verifyUser,(req, res, next) => {
+  .post(authenticate.verifyUser, authenticate.verifyAdmin,(req, res, next) => {
     res.statusCode = 403;
     res.end("POST operation not supported on /dishes/" + req.params.dishId);
   })
-  .put(authenticate.verifyUser,(req, res, next) => {
+  .put(authenticate.verifyUser, authenticate.verifyAdmin,(req, res, next) => {
     Dishes.findByIdAndUpdate(
       req.params.dishId,
       {
@@ -89,7 +89,7 @@ dishRouter
       )
       .catch((err) => next(err));
   })
-  .delete(authenticate.verifyUser,(req, res, next) => {
+  .delete(authenticate.verifyUser, authenticate.verifyAdmin,(req, res, next) => {
     Dishes.findByIdAndDelete(req.params.dishId)
       .then(
         (resp) => {
@@ -159,7 +159,7 @@ dishRouter
         " /comments"
     );
   })
-  .delete(authenticate.verifyUser,(req, res, next) => {
+  .delete(authenticate.verifyUser, authenticate.verifyAdmin,(req, res, next) => {
     Dishes.findById(req.params.dishId)
       .then((dish) => {
           if (dish != null) {
@@ -214,9 +214,10 @@ dishRouter
       + '/comments/'+ req.params.commentId );
   })
   .put(authenticate.verifyUser,(req, res, next) => {
+
     Dishes.findById(req.params.dishId)
     .then((dish) => {
-      if (dish != null && dish.comments.id(req.params.commentId)!=null) {
+      if (dish != null && dish.comments.id(req.params.commentId)!=null && dish.commentId(req.params.commentId).author.equals(req.user._id)) {
         if(req.body.rating){
           dish.comments.id(req.params.commentId).rating = req.body.rating;
         }
@@ -239,11 +240,16 @@ dishRouter
         err = new Error("Dish " + req.params.dishId + " not found ! ");
         err.statusCode = 404;
         return next(err);
-      } else {
-        err = new Error("Comment " + req.params.commentId + " not found ! ");
-        err.statusCode = 404;
-        return next(err);
-      }
+      } else if (dish.comments.id(req.params.commentId) == null) {
+        err = new Error('Comment ' + req.params.commentId + ' not found');
+        err.status = 404;
+        return next(err);            
+    }
+    else {
+        err = new Error('you are not authorized to update this comment!');
+        err.status = 403;
+        return next(err);  
+    }
       },
       (err) => next(errr))
       .catch((err) => next(err));
@@ -251,7 +257,7 @@ dishRouter
   .delete(authenticate.verifyUser,(req, res, next) => {
     Dishes.findById(req.params.dishId)
     .then((dish) => {
-      if (dish != null && dish.comments.id(req.params.commentId)!=null) {
+      if (dish != null && dish.comments.id(req.params.commentId)!=null   && dish.comments.id(req.params.commentId).author.equals(req.user._id)) {
             dish.comments.id(req.params.commentId).deleteOne();
           dish.save()
           .then((dish) => {
@@ -268,11 +274,16 @@ dishRouter
           err = new Error("Dish " + req.params.dishId + " not found ! ");
           err.statusCode = 404;
           return next(err);
-        } else {
-          err = new Error("Comment " + req.params.commentId + " not found ! ");
-          err.statusCode = 404;
-          return next(err);
-        }
+        }  else if (dish.comments.id(req.params.commentId) == null) {
+          err = new Error('Comment ' + req.params.commentId + ' not found');
+          err.status = 404;
+          return next(err);            
+      }
+      else {
+          err = new Error('you are not authorized to delete this comment!');
+          err.status = 403;
+          return next(err);  
+      }
       },
       (err) => next(errr))
       .catch((err) => next(err));
